chore: convert preprocessor to vite plugin
