+++ .changeset/curvy-flies-carry.md
@@ -1,0 +1,5 @@
+---
+'@sveltejs/enhanced-img': minor
+---
+
+feat: add support for targeting `enhanced\:img` in CSS
